requirements.txt
fastapi==0.124.0
Jinja2==3.1.6
openai==2.9.0
asyncpg==0.30.0
psycopg2-binary==2.9.11
psycopg==3.2.6
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette==0.50.0
Telethon==1.30.0
uvicorn==0.38.0
